+++ apps/api/src/app/endpoints/data-providers/ghostfolio/ghostfolio.service.ts
@@ -16,6 +16,7 @@ import {
   DERIVED_CURRENCIES
 } from '@ghostfolio/common/config';
 import { PROPERTY_DATA_SOURCES_GHOSTFOLIO_DATA_PROVIDER_MAX_REQUESTS } from '@ghostfolio/common/config';
+import { getAssetProfileIdentifier } from '@ghostfolio/common/helper';
 import {
   DataProviderGhostfolioAssetProfileResponse,
   DataProviderHistoricalResponse,
@@ -60,7 +61,13 @@ export class GhostfolioService {
               }
             ])
             .then(async (assetProfiles) => {
-              const assetProfile = assetProfiles[symbol];
+              const assetProfile =
+                assetProfiles[
+                  getAssetProfileIdentifier({
+                    symbol,
+                    dataSource: dataProviderService.getName()
+                  })
+                ];
               const dataSourceOrigin = DataSource.GHOSTFOLIO;
 
               if (assetProfile) {